requirements.txt
PyJWT~=2.8.0
decouple~=0.0.7
pydantic~=1.10.13
uvicorn~=0.15.0
fastapi~=0.70.0